all: Decouple pkg/util/test/assertions tests
