+++ pkg/util/test/assertions/elements_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package assertions
+package assertions_test
 
 import (
 	"reflect"
@@ -20,8 +20,9 @@ import (
 	"testing"
 
 	"github.com/smartystreets/assertions"
-	"github.com/smartystreets/assertions/should"
 	"go.thethings.network/lorawan-stack/v3/pkg/util/test"
+	. "go.thethings.network/lorawan-stack/v3/pkg/util/test/assertions"
+	"go.thethings.network/lorawan-stack/v3/pkg/util/test/assertions/should"
 )
 
 func TestShouldHaveSameElements(t *testing.T) {
